refactor: Clean up files
